pyproject.toml
[project]
name = "pronote2calendar"
version = "0.3.0"
description = "A simple tool to export your Pronote timetable to Google Calendar for easy access and integration with other apps."
authors = [
    {name = "Damien Chaib",email = "dchaib@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pronotepy (==2.14.5)",
    "google-api-python-client (==2.187.0)",
    "pydantic-settings (==2.12.0)",
]

[tool.isort]
profile = "black"

[tool.poetry]
packages = [{include = "pronote2calendar", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
test = [
    "pytest (==9.0.1)"
]
dev = [
    "black (==25.11.0)",
    "flake8 (==7.3.0)",
    "flake8-bugbear (==25.10.21)",
    "mypy (==1.18.2)",
    "isort (==7.0.0)"
]
